LF-3363 - Moves the object creation out of the controller, adds jsdoc from my boi GPT!
The upsert object is quite large and seemingly complex. It has been moved to the utility file to clean up the controller.

Hopefully the seperated out getManagementPlanTemplateGraph() can be used in standalone situations.

Jsdoc is a pain but also useful so I tried out GPT.
